feat: support use as a library
