+++ Makefile
@@ -1,0 +1,37 @@
+# Determine root directory
+ROOT_DIR=$(shell dirname $(realpath $(firstword $(MAKEFILE_LIST))))
+
+# Gather all .go files for use in dependencies below
+GO_FILES=$(shell find $(ROOT_DIR) -name '*.go')
+
+# Gather list of expected binaries
+BINARIES=$(shell cd $(ROOT_DIR)/cmd && ls -1 | grep -v ^common)
+
+# Extract Go module name from go.mod
+GOMODULE=$(shell grep ^module $(ROOT_DIR)/go.mod | awk '{ print $$2 }')
+
+# Set version strings based on git tag and current ref
+GO_LDFLAGS=-ldflags "-X '$(GOMODULE)/internal/version.Version=$(shell git describe --tags --exact-match 2>/dev/null)' -X '$(GOMODULE)/internal/version.CommitHash=$(shell git rev-parse --short HEAD)'"
+
+.PHONY: build mod-tidy clean test
+
+# Alias for building program binary
+build: $(BINARIES)
+
+mod-tidy:
+	# Needed to fetch new dependencies and add them to go.mod
+	go mod tidy
+
+clean:
+	rm -f $(BINARIES)
+
+test: mod-tidy
+	go test -v ./...
+
+# Build our program binaries
+# Depends on GO_FILES to determine when rebuild is needed
+$(BINARIES): mod-tidy $(GO_FILES)
+	go build \
+		$(GO_LDFLAGS) \
+		-o $(@) \
+		./cmd/$(@)
