+++ internal/version/version.go
@@ -1,0 +1,31 @@
+// Copyright 2023 Blink Labs, LLC.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package version
+
+import (
+	"fmt"
+)
+
+// These are populated at build time
+var Version string
+var CommitHash string
+
+func GetVersionString() string {
+	if Version != "" {
+		return fmt.Sprintf("%s (commit %s)", Version, CommitHash)
+	} else {
+		return fmt.Sprintf("devel (commit %s)", CommitHash)
+	}
+}
